requirements.txt
python-dotenv
streamlit-option-menu==0.3.12
pillow==10.2.0
pdf2image
pandas
leafmap
streamlit
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
python-pptx
streamlit-authenticator==0.1.5





